refactor: isolate featured categories card
